Add tokio::sync::Barrier to test_empty/test_full
Barrier ensures that the main test thread doesn't
empty (or fill) the queue before the spawned task
has a chance to start. The test still passes
without the Barrier, but could be passing only due
to the short-circuiting `self.if_empty()` or
`self.capacity() == self.len()` logic in the body
of `async fn empty()`/`full()`.